+++ CHANGELOG.md
@@ -14,8 +14,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Splash Screen
 - About
   - Credits
+- Toast messages and validation messages
 ### Improved
 - Some more Enums used in frontend
+- `console.error()` in some places into `toast.error(...,{description})`
 
 ## [1.0.0-2] - 2026-02-09 (v1 Beta #2)
 ### Added